+++ plugins/storage/fs/src/hopeit/fs_storage/partition.py
@@ -1,16 +1,17 @@
 """
 FS Storage plugin package module
 """
+
 from datetime import datetime, timezone
 
 from hopeit.dataobjects import DataObject
 
 
+def get_file_partition_key(partition_dateformat: str) -> str:
+    ts = datetime.now(tz=timezone.utc)
+    return ts.astimezone(timezone.utc).strftime(partition_dateformat.strip("/")) + "/"
+
+
 def get_partition_key(payload: DataObject, partition_dateformat: str) -> str:
-    ts = _partition_timestamp(payload)
-    return ts.strftime(partition_dateformat.strip('/')) + '/'
-
-
-def _partition_timestamp(payload: DataObject) -> datetime:
     ts = payload.event_ts() or datetime.now(tz=timezone.utc)  # type: ignore
-    return ts.astimezone(timezone.utc)
+    return ts.astimezone(timezone.utc).strftime(partition_dateformat.strip("/")) + "/"