+++ docs/source/release-notes.rst
@@ -1,6 +1,13 @@
 Release Notes
 =============
 
+Version 0.23.0
+______________
+- Plugin:
+  
+  - fs-storage: add `store_file`, `get_file`, `delete_files` and `list_files` methods
+
+
 Version 0.22.0
 ______________
 - Engine: